Add default raw client
