+++ conjure-runtime/src/body.rs
@@ -11,13 +11,19 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-use crate::BodyWriter;
+use crate::raw::BodyPart;
 use async_trait::async_trait;
-use bytes::Bytes;
+use bytes::{Bytes, BytesMut};
 use conjure_error::Error;
+use futures::channel::mpsc;
+use futures::{ready, SinkExt};
 use hyper::header::HeaderValue;
 use pin_project::pin_project;
+use std::io;
+use std::marker::PhantomPinned;
 use std::pin::Pin;
+use std::task::{Context, Poll};
+use tokio::io::{AsyncWrite, AsyncWriteExt};
 
 /// A request body.
 ///
@@ -189,5 +195,84 @@ where
             *this.needs_reset = false;
         }
         ok
+    }
+}
+
+/// The asynchronous writer passed to `Body::write`.
+#[pin_project]
+pub struct BodyWriter {
+    #[pin]
+    sender: mpsc::Sender<BodyPart>,
+    buf: BytesMut,
+    #[pin]
+    _p: PhantomPinned,
+}
+
+impl BodyWriter {
+    pub(crate) fn new(sender: mpsc::Sender<BodyPart>) -> BodyWriter {
+        BodyWriter {
+            sender,
+            buf: BytesMut::new(),
+            _p: PhantomPinned,
+        }
+    }
+
+    pub(crate) async fn finish(mut self: Pin<&mut Self>) -> io::Result<()> {
+        self.flush().await?;
+        self.project()
+            .sender
+            .send(BodyPart::Done)
+            .await
+            .map_err(|e| io::Error::new(io::ErrorKind::Other, e))?;
+        Ok(())
+    }
+
+    /// Writes a block of body bytes.
+    ///
+    /// Compared to the `AsyncWrite` implementation, this method avoids some copies if the caller already has the body
+    /// in `Bytes` objects.
+    pub async fn write_bytes(mut self: Pin<&mut Self>, bytes: Bytes) -> io::Result<()> {
+        self.flush().await?;
+        self.project()
+            .sender
+            .send(BodyPart::Chunk(bytes))
+            .await
+            .map_err(|e| io::Error::new(io::ErrorKind::Other, e))?;
+        Ok(())
+    }
+}
+
+impl AsyncWrite for BodyWriter {
+    fn poll_write(
+        mut self: Pin<&mut Self>,
+        cx: &mut Context<'_>,
+        buf: &[u8],
+    ) -> Poll<io::Result<usize>> {
+        if self.buf.len() > 4096 {
+            ready!(self.as_mut().poll_flush(cx))?;
+        }
+
+        self.project().buf.extend_from_slice(buf);
+        Poll::Ready(Ok(buf.len()))
+    }
+
+    fn poll_flush(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<io::Result<()>> {
+        let mut this = self.project();
+
+        if this.buf.is_empty() {
+            return Poll::Ready(Ok(()));
+        }
+
+        ready!(this.sender.poll_ready(cx)).map_err(|e| io::Error::new(io::ErrorKind::Other, e))?;
+        let chunk = this.buf.split().freeze();
+        this.sender
+            .start_send(BodyPart::Chunk(chunk))
+            .map_err(|e| io::Error::new(io::ErrorKind::Other, e))?;
+
+        Poll::Ready(Ok(()))
+    }
+
+    fn poll_shutdown(self: Pin<&mut Self>, _: &mut Context<'_>) -> Poll<io::Result<()>> {
+        Poll::Ready(Ok(()))
     }
 }
