+++ conjure-runtime/src/lib.rs
@@ -196,7 +196,6 @@ pub use crate::builder::*;
 pub use crate::client::*;
 pub use crate::client_factory::*;
 pub use crate::host_metrics::*;
-pub use crate::payload::*;
 pub use crate::request::*;
 pub use crate::response::*;
 pub use crate::user_agent::*;
@@ -211,7 +210,7 @@ mod client_factory;
 mod conjure;
 pub mod errors;
 mod host_metrics;
-mod payload;
+pub mod raw;
 mod request;
 mod response;
 mod service;
